[nextest-runner] move skip count formatters to another module
Rename the duration module to formatters.rs, and add the skip count formatter
to that.
